Card modulunde degisiklikler

+++ GCard/payment/models.py
@@ -3,7 +3,7 @@ from django.conf import settings
 
 class Card(models.Model):
      digits = models.CharField(max_length=8)
-     balance = models.PositiveIntegerField()
+     balance = models.PositiveSmallIntegerField()
 
 
 class Product(models.Model):
